feature: initial implementation of CategoriesRepository

+++ src/main/java/ml/meajudadev/api/v1/dto/ExpenseCategoryDto.java
@@ -3,8 +3,8 @@ package ml.meajudadev.api.v1.dto;
 import com.fasterxml.jackson.annotation.JsonProperty;
 
 public record ExpenseCategoryDto(
-        @JsonProperty("id") int id,
-        @JsonProperty("userId") int userId,
+        @JsonProperty("id") long id,
+        @JsonProperty("userId") long userId,
         @JsonProperty("name") String name,
         @JsonProperty("state") char state,
         @JsonProperty("type") char type
